Add optional callback at each iteration of rendering
