+++ lucent/modelzoo/__init__.py
@@ -1,2 +1,3 @@
 from torchvision.models import *  # noqa: disable=F401
+
 from lucent.modelzoo.inceptionv1 import InceptionV1 as inceptionv1  # noqa: disable=F401
